LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/css-color/system-color-consistency-expected.txt


FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a button assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 7 got 3
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a button
PASS Property color value 'ButtonText' resolves to the same color as text on a button
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a submit button assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 7 got 3
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a submit button
PASS Property color value 'ButtonText' resolves to the same color as text on a submit button
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a reset button assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 7 got 3
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a reset button
PASS Property color value 'ButtonText' resolves to the same color as text on a reset button
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a color picker assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 128 +/- 1, expected 128 but got 255
FAIL Property color value 'ButtonFace' resolves to the same color as the background-color of a color picker assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected 255 +/- 1, expected 255 but got 192
FAIL Property color value 'ButtonText' resolves to the same color as text on a color picker assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'CanvasText' has the same color as the color of the html element
PASS Property color value 'Field' resolves to the same color as the background-color of a text field
FAIL Property color value 'FieldText' resolves to the same color as text on a text field assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'Field' resolves to the same color as the background-color of a password field
FAIL Property color value 'FieldText' resolves to the same color as text on a password field assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'Field' resolves to the same color as the background-color of a email field
FAIL Property color value 'FieldText' resolves to the same color as text on a email field assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'Field' resolves to the same color as the background-color of a number field
FAIL Property color value 'FieldText' resolves to the same color as text on a number field assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'Field' resolves to the same color as the background-color of a date field
FAIL Property color value 'FieldText' resolves to the same color as text on a date field assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'Field' resolves to the same color as the background-color of a text area
FAIL Property color value 'FieldText' resolves to the same color as text on a text area assert_array_approx_equals: Numeric parameters are approximately equal. lengths differ, expected 3 got 4
PASS Property color value 'Mark' has the same color as the background-color of a mark element
PASS Property color value 'MarkText' has the same color as the color of a mark element
